refactor(sitemap): extract getSitemapUrls into web module
- add sitemap path into web module
- documentation of various functions used
- removed additional curly braces in web controller
- removed unnecessary import in file controller

re #400
